Rename `LanguagesDropdown` component
