+++ panel/src/components/Navigation/LanguagesDropdown.vue
