+++ panel/src/components/Navigation/index.js
@@ -10,7 +10,7 @@ import DropdownContent from "./DropdownContent.vue";
 import DropdownItem from "./DropdownItem.vue";
 import FileBrowser from "./FileBrowser.vue";
 import Link from "./Link.vue";
-import Languages from "./Languages.vue";
+import LanguagesDropdown from "./LanguagesDropdown.vue";
 import OptionsDropdown from "./OptionsDropdown.vue";
 import PageTree from "./PageTree.vue";
 import Pagination from "./Pagination.vue";
@@ -32,7 +32,7 @@ export default {
 		app.component("k-dropdown-content", DropdownContent);
 		app.component("k-dropdown-item", DropdownItem);
 		app.component("k-file-browser", FileBrowser);
-		app.component("k-languages-dropdown", Languages);
+		app.component("k-languages-dropdown", LanguagesDropdown);
 		app.component("k-link", Link);
 		app.component("k-options-dropdown", OptionsDropdown);
 		app.component("k-page-tree", PageTree);
